+++ app/__init__.py
@@ -1,6 +1,6 @@
 """Paquete de la aplicación TUSSAM API.
 
 Autor: 686f6c61 (https://github.com/686f6c61)
-Versión: 1.0.1
+Versión: 1.0.2
 Licencia: MIT
 """